goatfather (#2)
Effect Switcher library
GoatRoller use library and can edit patches
GoatFather code

+++ libraries/effectswitcher/Patch.h
@@ -8,16 +8,16 @@ class Patch {
 private:
     int patch_number;
     int eeprom_address;
-    int selected_effects;
+    byte selected_effects;
 
 public:
     Patch(int _patch_number, int _eeprom_address, int _selected_effects = 0, int _load_from_eeprom = true);
     ~Patch();
 
     void load();
-    void save(int _selected_effects);
+    void save(byte _selected_effects);
 
-    int get_selected_effects() const;
+    byte get_selected_effects() const;
 };
 
 #endif //GOATBOX_PATCH_H
